requirements.txt
python-magic
whoosh
nltk
